Extract operator traits to operators module
